refactor(cli): split profile command module [OPE-404]
